Import TMP + Cell Pick/Clicker
Setup a basic Cell Picker/Clicker so we can click and see coordinates of the selected cell.

Uses a mesh collider instead of rebuilding our mesh / generation math and working in reverse, idk its a bit iffy since either one is honestly kinda expensive. But the mesh collider is 1:1 sturdy and works on edges / visually clickable voxels and isn't going to be used/called honestly that much anyways if we setup tiles to mostly use animations and not the physics on the planet itself <3